+++ ExpandWorldPrefabs/InfoSelector.cs
@@ -8,31 +8,30 @@ namespace ExpandWorld.Prefab;
 
 public class InfoSelector
 {
-  public static Info? Select(ActionType type, ZDO zdo, string arg, Parameters parameters, ZDO? source)
+  public static Info? Select(ActionType type, ZDO zdo, string[] args, Parameters parameters, ZDO? source)
   {
     var infos = InfoManager.Select(type);
-    return SelectDefault(infos, zdo, arg, parameters, source) ?? SelectFallback(infos, zdo, arg, parameters, source);
+    return SelectDefault(infos, zdo, args, parameters, source) ?? SelectFallback(infos, zdo, args, parameters, source);
   }
-  private static Info? SelectDefault(PrefabInfo infos, ZDO zdo, string arg, Parameters parameters, ZDO? source)
+  private static Info? SelectDefault(PrefabInfo infos, ZDO zdo, string[] args, Parameters parameters, ZDO? source)
   {
     var prefab = zdo.m_prefab;
     if (!infos.TryGetValue(prefab, out var data)) return null;
-    return SelectInfo(data, zdo, arg, parameters, source);
+    return SelectInfo(data, zdo, args, parameters, source);
   }
-  private static Info? SelectFallback(PrefabInfo infos, ZDO zdo, string arg, Parameters parameters, ZDO? source)
+  private static Info? SelectFallback(PrefabInfo infos, ZDO zdo, string[] args, Parameters parameters, ZDO? source)
   {
     var prefab = zdo.m_prefab;
     if (!infos.TryGetFallbackValue(prefab, out var data)) return null;
-    return SelectInfo(data, zdo, arg, parameters, source);
+    return SelectInfo(data, zdo, args, parameters, source);
   }
-  private static Info? SelectInfo(List<Info> data, ZDO zdo, string arg, Parameters parameters, ZDO? source)
+  private static Info? SelectInfo(List<Info> data, ZDO zdo, string[] args, Parameters parameters, ZDO? source)
   {
     if (data.Count == 0) return null;
     var pos = zdo.m_position;
     var biome = WorldGenerator.instance.GetBiome(pos);
     var distance = Utils.LengthXZ(pos);
     var day = EnvMan.IsDay();
-    var args = arg.Split(' ');
     var waterY = pos.y - ZoneSystem.instance.m_waterLevel;
     var linq = data
       .Where(d => CheckArgs(d, args))
@@ -216,17 +215,16 @@ public class InfoSelector
     Random.state = state;
     return env.m_name.ToLower();
   }
-  public static Info? SelectGlobal(ActionType type, string arg, Parameters parameters, Vector3 pos, bool remove)
+  public static Info? SelectGlobal(ActionType type, string[] args, Parameters parameters, Vector3 pos, bool remove)
   {
     var infos = InfoManager.SelectGlobal(type);
-    return SelectGlobalInfo(infos.Info, arg, parameters, pos, remove) ?? SelectGlobalInfo(infos.Fallback, arg, parameters, pos, remove);
+    return SelectGlobalInfo(infos.Info, args, parameters, pos, remove) ?? SelectGlobalInfo(infos.Fallback, args, parameters, pos, remove);
   }
 
-  private static Info? SelectGlobalInfo(List<Info> data, string arg, Parameters parameters, Vector3 pos, bool remove)
+  private static Info? SelectGlobalInfo(List<Info> data, string[] args, Parameters parameters, Vector3 pos, bool remove)
   {
     if (data.Count == 0) return null;
     var day = EnvMan.IsDay();
-    var args = arg.Split(' ');
     var distance = Utils.LengthXZ(pos);
     var waterY = pos.y - ZoneSystem.instance.m_waterLevel;
     var linq = data